Minor bugs fixes still remaining

+++ backend/src/main/java/com/example/gov_scheme_backend/dto/request/disbursement/StageDto.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -12,6 +13,6 @@ import java.time.LocalDate;
 public class StageDto {
     private Integer stageNumber;
     private String milestoneName;
-    private Double amountToRelease;
+    private BigDecimal amountToRelease;
     private LocalDate dueDate;
 }
